% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pudutest-package.R
\docType{package}
\name{pudutest-package}
\alias{pudutest}
\alias{pudutest-package}
\title{pudutest: Testing pudu}
\description{
Testing pudu. I run this within a tiny package to test the compiled code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pachadotdev/pudu}
  \item Report bugs at \url{https://github.com/pachadotdev/pudu/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
